Fix remaining u64/usize type mismatches after syscall argument change
- usrlib: sleep now passes SyscallNum as usize and ms as usize
- usrlib: wait_future and is_future_completed use handle.pack() instead
  of the manual u64 shift expressions
- kernel: TimeFuture::new receives arg1 as u64 since it stores a
  genuine 64-bit millisecond duration

+++ kernel/src/syscall.rs
@@ -21,7 +21,7 @@ pub fn handle_syscall(num: usize, arg1: usize, arg2: usize, arg3: usize) -> usiz
             0
         }
         Ok(SyscallNum::Sleep) => {
-            let future = Box::new(TimeFuture::new(arg1));
+            let future = Box::new(TimeFuture::new(arg1 as u64));
             let handle = services().future_registry
                 .borrow_mut()
                 .register(future)
